Finished porting the cuts.
Now the cuts can be done with the cutter class, which returns a cutmask which
is 0 if the event has been selected. Every cut has it's own class which is
customized for the cut. It has been verified that the cutter cuts the same
events as the treereader did before.