Changed the cuts.
Now there is a list of standard cuts which are applied for every histogram.
These are just the cuts in the same order as before, so now every histogram is
available at every step in the cut procedure. The order of cuts was kept from
the old way of doing things.