Arbitrary number of final state particles and cosmetics.
bannt can now work with an arbitrary number of particles, defined in
constants.hpp (which has been renamed from constants.h). This required some
restructuring of the Event and Data classes and changed the way the
BranchAddresses are set, but had surprisingly little impact on the rest of the
program. Also made the Plotter and Event singletons, as they are used like this
anyway.