Storing more CEDAR information.
In addition to the likelihood method, now storing also the "old" majority
method for the CEDAR PID in the TTree. To get a better handle on the effects
observed when using the likelihood method, the theta angle in the CEDAR is
saved as well. For this, a plot was already added while the majority method
still needs to be implemented as cut.