antok: Further on the initializer.
The variables in the tree are now defined by the config file: one can specify
branches which are present in the tree in the config file. There is a
distinction between data types and branches which are there once for every
particle. The data are stored in maps in the Data object and can be accessed
with their string.

This is an intermediate commit: there is some debug output in and the values in
the map are not used anywhere in the rest of the code yet.