antok: Intermediate commit.
This commit is far from functional, but I want to get the working directory
cleaned up a bit, as I have to change the changes yet again and want some
fall-back point.

So far, an ObjectManager class was introduced which holds all the objects
needed. This was done to avoid passing around objects (which are all singletons
anyway) too much. Also introduced function classes which do calculations for
every event. Started on the initialization of the Event class in the
Initializer, but this turns out to be quite extensive and the approach taken so
far might not work out. Will see...