antok: More functions and restructuring.
Added more functions to the Initializer. While doing that, the Initializer was
restructured so that every function object gets initialized in its own
functions, which cleans up the lengthy process somewhat. Still, there is a lot
of code repetition, which is not nice. Will try to fix that, but that needs
more restructuring which I do not want to start before committing.