antok: Implemented all the stuff Event calculated.
Now all the stuff which was calculated in the Event::update() can be done via
the configuration file. The initializer is doing quite some work to register
all the functions correctly, but the main part is handling of errors and
indices. With a bit of restructuring, I was able to cut down on code in the
methods actually registering the Functions, however, the Initializer is still
by far the most heavy piece of code in the whole program.