antok: Some cosmetics and first RangedCut.
Some minor cosmetic changes here and there. The main change is the introduction
of a RangedCut class in cut.hpp, inheriting form the usual Cut class. This
should already cover most of the cuts which are used at the moment in the
event selection. The hard thing will be to figure out how to do correct
bookkeeping of the cuts.