antok: Further on the Initializer.
Multiple improvements to the way functions are registered: some pieces of code
were moved to the argumentHandler, further reducing code duplication in the
code registering the Functions. Also, YAML's as<> functions were made exception
safe, so that an error preventing conversion will not cause a crash without any
useful hint as to were the error originated from. Also added an "abs" and a
"diff" function.

In parallel, most of the cuts which were used before were added to the
configuration file. This was only just started though, as it prompted the
improvements to the initializer mentioned above.

As preparation for the cut machinery, the object manager now also holds a
pointer to the output file.