antok: Reshuffling.
Created a new namespace antok::generators, which now contains the functions
producing the antok::Function*s. This makes for a better organization of code,
but some reshuffling and renaming of utility functions remains to be done.

Also introduced the namespace antok::functions, in which all the specialized
functions deriving from the Function class (which remains in namespace antok)
are located. This way, the antok namespace does not get too cluttered.