antok: Clean-up.
Removed or commented out all code which was left from the old-way of doing
things. Were it is commented it still carries information which was not taken
over to the YAML-style way.

Also changed Data from header only to a header/source implementation, as it was
gathering quite some functions.