antok: First parsing of cut options.
The range cut is now correctly translated to a RangeCut object, with the usual
safety checks and this seems to work. The framework for this is still crude and
does not implement different CutTrains yet, but this should only be a question
of organization.

The cut objects for EqualityCuts and CutGroups are already there, but cuts of
these types are not yet translated while the TriggerMaskCut is missing
altogether for now.