antok: More on cuts and restructuring.
Redid the way the cuts are handled by introducing the CutTrains entry in the
config file. This way, multiple event selections should be possible by
specifying multiple CutTrains.

Also added a function for calculating the energy of a TLorentzVector and redid
a lot of Data's member functions to be templated.

Finally moved the getYAML* functions to their own file, as this seems to be the
only choice which makes sense.