antok: First file management and reversed cut logic.
Added support for multiple output TTrees and introduced a general framework to
handle root TObjects which get saved in the output file. For every CutTrain, a
directory is created in the output file and one can choose if the CutTrain gets
saved to a TTree. Also added some first code to initialize the Plotter, but
nothing substantial yet.

In addition, the logic of the cuts was changed: if the cut was survived, the
output value is now 'true' (before it was 'false'). This should make the
cutmasks more logical ('1' to enable a cut, '0' to disable).