antok: Revised the Plot.
Adjusted the Plot's functions to the new cutmask logic and structure. In the
course of this, several new functions were introduced in the Cutter (mostly
getters to retrieve information). The initializePlotter function is still a
construction site and in its current state, execution fails with an assertion
error. Will work on that.