antok: Basic plotting works.
The Basic plots are generated for all CutTrains, 1D and 2D. This necessitated
some restructuring of the Plot to allow for double and int plots to be drawn.
Also moved the basic work into a new method in the ObjectManager. The output
TTrees are not yet being filled, that will be the next step.