antok: Bugfixes.
Fixed a few bugs that were found during crosschecking with the version tagged
as "event_selection_run4" (d0c4ac1b4e21b1f994354d51e9623d1b6e24ee5a). The worst
bug was that the sum2 (squared sum + square root) was broken, the others were
just minor problems with root directory management and some memory management.

During the comparison with the old version, it was discovered that the RPD
helper is giving out NaN for the proton mass sometimes. As the comparison logic
changed (true is now accepted, was rejected in the old version) and as
comparisons with NaN are always false, this makes a difference in the number of
events which survive the RPD proton mass cut and the planarity cut. However,
the new version should be the one to be trusted, as it requires a true value
for the event to survive the cut, rejecting all NaNs.