antok: Waterfall cutmasks produced by the cutter.
The waterfall cutmasks, which were produced by the Initializer when
initializing the cutter until now, are now produced by the cutter itself on
demand only. The cutmasks are cached, so they are only produced by the first
call to the function, which means it should only be called after all cuts have
been added.