antok: Introduced GlobalPlotOptions.
Introduced the GlobalPlotOptions, with which a number of plotting features can
be controlled. Most of it is implemented, however, the StatisticsHistogram
option is still missing.

Also found a problem in the YAML parts: if there the existence of a key is
tested and the node in question is not a map, the test throws an exception.
This is a problem as then there is no indication of where exception came from.
As this is used throughout the code, a fix will be quite some work.