antok: Fixed the key-checking problem.
There is now a function in the antok::YAMLUtils namespace which checks if a
node has a key and catches the exception which is thrown if the node is a
scalar. I tried to remove all the old checks and replace them with the new
ones, but as this mechanism was used quite extensively, it is possible that a
few slipped through.