antok: Implemented CustomPlots and a bugfix.
The CustomPlots entry in the configfile for single individual plots now is
implemented and seems to work correctly.

Also found a bug in the cutter: when getting the cutmasks for single plots
on/off, the order was not the same as specified in the config file. This is
fixed now.