antok: Fill plots several times.
In order to fill a plot with data from all outgoing particles of one event, it
is now possible to put the "Indices" option to a plot. This will cause the plot
to be filled with the variables specified by their name plus the appended
index for all indices for every event. It work for 1D and 2D plots.