antok: Cosmetics.
Harmonized some formatting: there should now be a empty line at the end of
every file and the #include statements should now follow roughly the same
structure. Also improved some of the #includes and removed all the commented
code which was left from before the change to the yaml config file.