antok: More efficient plotting.
When instantiating the histograms, check if that every histogram is only
produced once for every cutmask. Then, when the program has finished, add the
histograms which were skipped by cloning the ones which were filled.

Unfortunately, this messes up the order in which the histograms appear in the
root file, which is very inconvenient. Will investigate how this can be fixed.