antok: Fix the problems with the ordering of plots.
Remember the order in which the plots are generated in the template plot. Pass
this information to the initializer, who saves it the object manager. Together
with the registration of objects which have to be copied, the object manager
will either "touch" (remove and re-append) the histogram or clone it from its
"parent".

Might have to be rethought: it might be better to put all histograms (without
the duplicates) in a temporary directory, then generate the output directories
and close them one by one. This could maybe reduce the memory footprint at the
end of the program.