antok: Improve memory usage for plots.
In order to be able to close directories as soon as all histograms in them have
been written, write all plots to a temporary directory first. At the end, copy
them to their nominal location and close the directory as soon as all plots
belonging in it have been written.