antok: Add new cutTrain for MC and fix bug.
Add a new cutTrain for the phase space Monte Carlo configuration file which
accepts all events. Make it pertinent so that the events are saved into a tree.

Also fix a small problem in the finalize step of the object manager: write the
output trees before writing the histograms to make sure that the directories
the trees go in are not closed before the trees have been written.