antok: Add tool to generate beam histogram.
Add a new tool which takes the data tree from antok and generates a 5d sparse
histogram with the variables needed for the beam simulation (vertex position
and beam momentum). This is quite a hack, several things are hard coded, but so
far it only was used as a test case, so it might be polished a bit more in the
future.