antok: Try a naive solution to get the beamfile.
For every event, loop over all other events to find the 10 events which are
closest and determine the sigma that way. This cannot work, it takes on the
order of 10 seconds to do that for one event (with 10'000'000 in total) and
scales with n*n. This commit will be reverted.