antok: Add better solution to produce beamfile.
Add a calculation of the sigmas which ensures at least 10 events are used to
determine the average distance between events. For this, use a 5D histogram and
increase the number of bins used for the calculation until they contain at
least 10 bins.

There is a lot of debug code and commented code still in, which will be removed
in the next commit.