antok: Change method to create beamfile.
Revise the tool to create a beamfile: as suggested by Jan Friedrich, use an
adaptive binning to determine the sigmas. To get the bins, split the
phase-space in two along one axis and then split the resulting two sub-spaces
in two along the next axis. Continue until the number of events in the subspace
are below a certain limit. Then, calculate the sigmas with the dimensions of
the resulting subspaces and the entries therein. This seems to be much faster
than the cube method and will probably yield more reliable results.

The old code is still in, but commented out, will clean that up in the next few
commits.

To get the linker to work, adjust the FindROOT.cmake macro. It now links
against the EVE libraries. Also fix a bug, which probably would have lead to
problems once root-config would actually output auxlibs.