antok: Investigate memory use of generateBeamfile.
The generateBeamfile tool uses a lot of memory, add some code to reduce the
footprint and some debug code to investigate the issue. However, it turned out
that the problem lies in the overhead of the TTree class, which uses about
11 KB of memory even when only one entry is saved. There will be some
restructuring in the next commits to fix this issue.