antok: Fix minor bug.
When initializing the Data object in the Initializer, the Data was instantiated
before it was checked that a config file had been read. If the config file was
not read and initAll was called, the Initializer was left in a state where it
could not be initialized anymore (because of the already existing Data object)
even if the config file was read after the initial call. Fix this by reordering
the consistency checks.