antok: Introducing pytok.
Add bindings for python for the constants and basic calculations from antok.
This is done as a preparation to harmonize the python scripts and library bits
which have been developed over time and put them all into one central python
module. The deeper inner workings of the antok library will not be exposed in
python, just the basics which could be useful in scripts.