antok + MC Scripts: Major restructuring.
Move a lot of files around: All the files in the mcBatchLib go into the pytok
module. The job scripts go into the helperScript subdirectory. Rename all the
.cc files to .cxx in the pytok bindings.

This badly breaks the whole Monte Carlo Script system for the moment. In order
to have clean renames, none of the files has been adjusted to their new
locations yet, that will follow in the next few commits.