antok: Add generateBeamfileTest.
Add a second version of generateBeamFile for testing purposes. This version,
instead of calculating the relevant values stored in the beamfile, just loads
the data from a tree, making it possible to load "artificial" beamfiles (i.e.
beamfile trees which have been generated randomly in one way or another). This
file will probably be removed again once testing is complete.