antok: Work on the beamfile generator.
Modify generateBeamfile.cxx so that it can easily be switched between
processing data with different numbers of final state particles. Also change
back to the way the sigmas were calculated earlier (dividing edge length of the
bin by the number of entries in it), as it seems to be more correct. Fix the
problems from before by making sure that the bins were divided an equal number
of times in all dimensions.

In the beamfile_filtering.yaml, add the NoCut. It is not used but it is
convenient to have it in there for quick processing.