antok: Revise beamfileGenerator.
Revise the 5 dimensional bin and move it to its own file. It now collects a lot
more information: it knows which bins are its neighbors and if it is a bin on
the edge of the phase space volume. It also does some calculations which were
done in the getAdaptiveBins() function before. Change the memory management of
the 5d bins, they are now all held by shared_ptrs.

In addition, add some cosmetic changes here and there.

In a next step, the calculation of the sigmas will be changed to take the
additional information into account.