antok: Revise sigma calculation yet again.
Revise the sigma calculation again from scratch: the method implemented now is
a mixture between the two tried before: the idea is to calculate a
hyperrectangle for every event which has a volume equal to the mean volume per
event and the same ratios of edge lengths as the bin the event is located in.
This should make the sigma calculation independent of bin size and content.

Also add a debug flag to the 5d bin class and make several smaller mainly
cosmetic changes. Harmonize generateBeamfile and generateBeamfileTest.