antok: Add some feature to the beamfile generator.
Add the possibility to add indentation to print() of the five dimensional bin.
This allows on the one hand to print the neighbors of a bin when printing the
bin itself (implement this as optional) as the neighbors can be indented. On
the other hand, one can show the depth in the adaptive binning recursion by
indenting the printing proportional to the depth of the recursion.

Also add some more information to the output TTree, it being the bin volume,
the "edgeity" (the number of coordinates the bin is at the lower or upper edge)
and the number of neighbors.

Simplify the areWeNeighbors function quite significantly, with it producing the
same results. The only thing changes is that bins which have overlap (they
share some common volume) are now recognized as neighbors which was not the
case before.