antok: Move stuff around in the beamfileGenerator.
Move the events in the bins into the fiveDimBin class itself. This makes a lot
of interfaces a lot more elegant, however, a lot of code had to be touched to
do it. Also move the "edgeity" calculation to the 5d bin and prepare for a
different sigma calculation in case of edge bins.