antok: Sigma calculation for edge bins.
Add sigma calculation for bins on the edge of the phase space: if the bin in
question has neighboring bins which are not on the edge, take the mean of their
sigmas as the sigma. In case there are no such neighbors, take the closet
events for every event and calculate the sigma by looking at the volume of the
5-ball they are in.

There is some old commented code still in the sigma calculation, will be
deleted in the next commit.