antok: Revise the sigma calculation again.
In the beamfile generator, revise the sigma calculation: change the way the
scaling is done for edge-edge bins (scale everything before starting the
business with the spheres). Also move to lower and upper sigma which are chosen
such that they do not exceed the bin boundaries.

This still leads to strange results, because the averaging of the sigmas in
edge bins now already uses the capped sigmas, which does not make sense. This
will be fixed in the next commit.