antok: Go back to single sigmas for the beamfile.
After some testing, go back to only one sigma per event per dimension in the
beamfile instead of the lower/upper-pair. For the single sigma, either take it
as calculated or limit it by comparing it with the distance of the event to the
bin boundaries. This is switchable with an option and is off by default.