antok: Add function.
Add a function to get the attributes as doubles from the TLorentzVector. This
cover getting lorentzVec.X(), .Y(), .Z(), .Phi() and .Theta(). Also put an
example to default.yaml and take the opportunity to adjust the path to
the input root tree and statistics histogram to change done in the
UserEvent.