antok: Add function to get the RPD's expected hits.
Add a function which calculates the expected hits in the RPD rings A and B.
Implement it in a separate function in basic_calcs so that coordinate
transformations can be applied which might be needed to accommodate the
position of the RPD in the experiment. For the moment, only the interfaces for
the coordinate transformation are there, but the function will throw an
exception if it is requested to actually do anything.