antok: Add transformation to the new function.
Add the actual transformation to the function calculating the RPD hitmaps.
There is a lot of debug code still in and the function might throw an unhandled
exception in some cases. The debug code will be removed in the next revision
and I will have to think about how to handle the exception.