antok: Add new method to get RPD delta phi.
Add a new method to obtain the RPD delta phi: similarly to the hitmap,
calculate from the beam and X system vectors the expected vector of the recoil
proton. Then compare the phi angle of the predicted proton with the phi angle
from obtained from the RPD, all in the lab system. For the Monte Carlo Truth,
this angle is exactly zero for all events, which makes it very easy to get the
probability distributions from the reconstructed events, as they are equal to
the distributions of the reconstructed events.