antok: Create antok::utils namespace.
Create a new namespace antok::utils and move the function solving quadratic
equations there. It is currently only used in basic_calcs' function to get the
RPD delta phi with the prediction method, but it will be useful in the
fitRPDResolutions as well.

Also rename the RPD ring diameters to radius, as that is the actual meaning of
these variables.