antok: Calculate length in target.
Add a new class in the fitRPDResolutions file which will be used to bin in t,
m(5 Pi) and l (length of proton path in target). Add a static function to
calculate l and add some dummy functions. There is quite some debug code of the
l calculation, will be cleaned up in next commit.