antok: Work on fitRPDResolutions tool.
Do a lot of things in the fitRPDResolutions tool, it can now fit also in t and
m bins (the binning in l was dropped, as it was unnecessary) and it saves
graphs of the combined fit parameters of every t/m bin. At the moment, the
parameters which are already available through the RpdHelperHelper are used and
fixed while fitting the distributions.