antok: Handle strange proton phi angles correctly.
Change the handling of the proton phi angle in a way that now there is a list
of valid angles the proton coming from the rpd is allowed to have. If the angle
is not among the valid ones, the likelihood given back is 0. This seems to do
the trick, now the cut seems to work reasonably.