antok: Add support for vector<double>.
Add the possibility to read std::vector<double> from the input file in the
treereader. Also add support to read TVector3 and TLorentzVector branches
directly from the input file (they were supported as internal data structures,
but were not added to the data types which could be read from the input).