antok: Create user area.
Create a new "user" subdirectory, where user specific code can be put. The idea
is that quantities calculated for specific analyses go in there. In the spirit
of this, move the RPD related functions to the newly created user/kbicker*
files. To facilitate adding new user code, there are also two template files
which can be used to generate the base files for new users.