antok: Fix std::vector<double> TTree branches.
Fix the reading of branches of the input TTree of type std::vector<double>.
There was a typo in the name of the tag for variables of this type, fix it.
Also, do not save std::vector<double>s but pointers to that, as that harmonizes
better with the root philosophy.