antok: Allow plotting std::vector variables.
Add handling of std::vector type variables in the plotting code. This makes it
possible to fill a plot with variables of type std::vector for every event,
meaning every number in the std::vector is filled into the plot. As should be
the case for std::vectors, this works it has different lengths for every event.
Also enable this for plots with the "Indices" tag, meaning that vectors present
for multiple particles (or multiple times for whatever other reason) can be
filled into the same plot as well.