Bugfix: Wrong output tree when merging more files
When merging more files without an own output tree definition,
Antok simply makes an Clone of the input tree of the first file.
If the first input files is closed, the branch addresses of the ouput trees are
nonsens.

Now when updating the input file, the branch addresses of all
output trees are set to thoes of the current input tree.