[getTs] Implement correct formula for |t|min
So far the approximation for |t|min was used which is correct up to
approx. 3%. 
Now the target mass can be given such that the exact value is
calculated. 
However, the numerical uncertainty of the exact value is very big for
some extreme kinematics (in the same order of magnitude as the
approximation).