configfile: Implement includes from other config files in configfile
Implement an include statement '!include' in the yaml config files,
which allows to include other config files.
This could only be implemented in pyyaml, thus a config file with an
include cannot be used with treereader.
The runParallel script now reads the config file with includes, resolves
the includes and writes a new config file with all information in one
file.
A new wrapper script for the treereader ('run') is added, which creates
a temporary config file for treereader and does the same as runParallel.
Additionally, yamltok, our patched version of pyyaml, does no longer
resolve yes/no/on/off to boolean variables, but keeps them as strings.