Fixed hasNodeKey
The YAML::Node operator[] cannot be returned as bool and therefore
a dummy access by key is performed - non-existing nodes will cause
an exception that is handled in the try-catch block. If the access
operation is successful true is returned. In case of an exception
the false statement is returned.